% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_config.R
\name{rclone_config_delete}
\alias{rclone_config_delete}
\title{Delete an rclone remote}
\usage{
rclone_config_delete(name)
}
\arguments{
\item{name}{Remote name to delete.}
}
\value{
Invisibly, the result of the rclone command.
}
\description{
Delete an rclone remote
}
\examples{
\dontrun{
rclone_config_delete("myremote")
}
}
