% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_config.R
\name{rclone_config_set}
\alias{rclone_config_set}
\title{Set individual parameters on an existing remote}
\usage{
rclone_config_set(name, ...)
}
\arguments{
\item{name}{Remote name.}

\item{...}{Named parameters to set (key = value).}
}
\value{
Invisibly, the result of the rclone command.
}
\description{
Set individual parameters on an existing remote
}
\examples{
\dontrun{
rclone_config_set("myremote", region = "us-west-2")
}
}
