% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_config.R
\name{rclone_config_show}
\alias{rclone_config_show}
\title{Show rclone configuration}
\usage{
rclone_config_show(name = NULL)
}
\arguments{
\item{name}{Optional remote name. If omitted, shows all remotes.}
}
\value{
Character string with the configuration (invisibly). Also prints
to the console.
}
\description{
Show rclone configuration
}
