% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_ops.R
\name{rclone_delete}
\alias{rclone_delete}
\title{Delete files at a remote path}
\usage{
rclone_delete(path, ...)
}
\arguments{
\item{path}{Remote path to delete.}

\item{...}{Additional rclone flags.}
}
\value{
Invisibly, the rclone result list.
}
\description{
Deletes the files (not directories) at the given path. To remove a directory
and all its contents use \code{\link[=rclone_purge]{rclone_purge()}}.
}
\examples{
\dontshow{if (rclone_available()) withAutoprint(\{ # examplesIf}
d <- tempfile()
dir.create(d)
writeLines("hi", file.path(d, "f.txt"))
rclone_delete(d)
\dontshow{\}) # examplesIf}
}
