% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_ops.R
\name{rclone_link}
\alias{rclone_link}
\title{Generate a public link for a remote file}
\usage{
rclone_link(path)
}
\arguments{
\item{path}{Remote path to the file.}
}
\value{
Character string with the public URL.
}
\description{
Not all remotes support this operation.
}
\examples{
\dontrun{
rclone_link("myremote:bucket/file.csv")
}
}
