% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_config.R
\name{rclone_listremotes}
\alias{rclone_listremotes}
\title{List configured remotes}
\usage{
rclone_listremotes()
}
\value{
Character vector of remote names.
}
\description{
List configured remotes
}
\examples{
\dontshow{if (rclone_available()) withAutoprint(\{ # examplesIf}
rclone_listremotes()
\dontshow{\}) # examplesIf}
}
