% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_ops.R
\name{rclone_mkdir}
\alias{rclone_mkdir}
\title{Create a directory or bucket}
\usage{
rclone_mkdir(path)
}
\arguments{
\item{path}{Remote path to create.}
}
\value{
Invisibly, the rclone result list.
}
\description{
Create a directory or bucket
}
