% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_ops.R
\name{rclone_rmdir}
\alias{rclone_rmdir}
\title{Remove an empty directory}
\usage{
rclone_rmdir(path)
}
\arguments{
\item{path}{Remote path to the directory.}
}
\value{
Invisibly, the rclone result list.
}
\description{
Remove an empty directory
}
