% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_ls.R
\name{rclone_size}
\alias{rclone_size}
\title{Get total size of a remote path}
\usage{
rclone_size(path)
}
\arguments{
\item{path}{Remote path.}
}
\value{
A named list with \code{count} (number of objects) and \code{bytes} (total
size in bytes).
}
\description{
Get total size of a remote path
}
\examples{
\dontshow{if (rclone_available()) withAutoprint(\{ # examplesIf}
rclone_size(tempdir())
\dontshow{\}) # examplesIf}
}
