% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_ls.R
\name{rclone_stat}
\alias{rclone_stat}
\title{Get metadata for a single object}
\usage{
rclone_stat(path)
}
\arguments{
\item{path}{Remote path to a single object.}
}
\value{
A one-row data frame with object metadata, or an error if not found.
}
\description{
Get metadata for a single object
}
