% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_ops.R
\name{rclone_version}
\alias{rclone_version}
\title{Show rclone version information}
\usage{
rclone_version()
}
\value{
Invisibly, a character string with the version output.
}
\description{
Show rclone version information
}
